[Feature] reafactor map configuration
