--- lib/blocs/configuration/models/transport_configuration.dart
@@ -1,5 +1,0 @@
-class TransportConfiguration {
-  final bool showTransportMarker;
-
-  const TransportConfiguration({this.showTransportMarker = true});
-}
