+++ lib/pages/home/plan_map/plan_map.dart
@@ -3,12 +3,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:trufi_core/blocs/configuration/configuration_cubit.dart';
-import 'package:trufi_core/blocs/configuration/models/transport_configuration.dart';
+import 'package:trufi_core/blocs/configuration/models/itinerary_creator.dart';
+import 'package:trufi_core/blocs/configuration/models/map_configuration.dart';
 import 'package:trufi_core/composite_subscription.dart';
 import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
-import 'package:trufi_core/models/enums/enums_plan/enums_plan.dart';
-import 'package:trufi_core/models/enums/enums_plan/icons/other_icons.dart';
-import 'package:trufi_core/models/markers/marker_configuration.dart';
 import 'package:trufi_core/pages/home/plan_map/plan.dart';
 import 'package:trufi_core/trufi_app.dart';
 import 'package:trufi_core/widgets/map/buttons/crop_button.dart';
@@ -28,16 +26,14 @@ class PlanMapPage extends StatefulWidget {
     Key key,
     @required this.customOverlayWidget,
     @required this.customBetweenFabWidget,
-    @required this.markerConfiguration,
-    @required this.transportConfiguration,
+    @required this.mapConfiguration,
     this.planPageController,
   }) : super(key: key);
 
   final PlanPageController planPageController;
   final LocaleWidgetBuilder customOverlayWidget;
   final WidgetBuilder customBetweenFabWidget;
-  final MarkerConfiguration markerConfiguration;
-  final TransportConfiguration transportConfiguration;
+  final MapConfiguration mapConfiguration;
 
   @override
   PlanMapPageState createState() => PlanMapPageState();
@@ -57,8 +53,7 @@ class PlanMapPageState extends State<PlanMapPage>
     _data = PlanMapPageStateData(
       plan: widget.planPageController.plan,
       onItineraryTap: widget.planPageController.inSelectedItinerary.add,
-      markerConfiguration: widget.markerConfiguration,
-      transportConfiguration: widget.transportConfiguration,
+      mapConfiguration: widget.mapConfiguration,
     );
     _subscriptions.add(
       _trufiMapController.outMapReady.listen((_) {
@@ -216,16 +211,15 @@ class PlanMapPageStateData {
   PlanMapPageStateData({
     @required this.plan,
     @required this.onItineraryTap,
-    @required this.markerConfiguration,
-    @required this.transportConfiguration,
+    @required this.mapConfiguration,
   }) {
     if (plan != null) {
       if (plan.from != null) {
-        _fromMarker = markerConfiguration
+        _fromMarker = mapConfiguration.markersConfiguration
             .buildFromMarker(createLatLngWithPlanLocation(plan.from));
       }
       if (plan.to != null) {
-        _toMarker = markerConfiguration
+        _toMarker = mapConfiguration.markersConfiguration
             .buildToMarker(createLatLngWithPlanLocation(plan.to));
       }
     }
@@ -233,8 +227,7 @@ class PlanMapPageStateData {
 
   final PlanEntity plan;
   final ValueChanged<PlanItinerary> onItineraryTap;
-  final MarkerConfiguration markerConfiguration;
-  final TransportConfiguration transportConfiguration;
+  final MapConfiguration mapConfiguration;
 
   final _itineraries = <PlanItinerary, List<PolylineWithMarkers>>{};
   final _unselectedMarkers = <Marker>[];
@@ -306,14 +299,11 @@ class PlanMapPageStateData {
     if (_toMarker != null) {
       _selectedBounds.extend(_toMarker.point);
     }
-    _itineraries.addAll(
-      _createItineraries(
-        plan: plan,
-        selectedItinerary: _selectedItinerary,
-        onTap: onItineraryTap,
-        showTransportMarker: transportConfiguration.showTransportMarker,
-      ),
-    );
+    _itineraries.addAll(mapConfiguration.itinararyCreator.buildItineraries(
+      plan: plan,
+      selectedItinerary: _selectedItinerary,
+      onTap: onItineraryTap,
+    ));
     _itineraries.forEach((itinerary, polylinesWithMarker) {
       final bool isSelected = itinerary == _selectedItinerary;
       for (final polylineWithMarker in polylinesWithMarker) {
@@ -363,90 +353,4 @@ class PlanMapPageStateData {
       orElse: () => null,
     );
   }
-
-  Map<PlanItinerary, List<PolylineWithMarkers>> _createItineraries({
-    @required PlanEntity plan,
-    @required PlanItinerary selectedItinerary,
-    @required Function(PlanItinerary) onTap,
-    bool showTransportMarker = true,
-  }) {
-    final Map<PlanItinerary, List<PolylineWithMarkers>> itineraries = {};
-    if (plan != null) {
-      for (final itinerary in plan.itineraries) {
-        final List<Marker> markers = [];
-        final List<PolylineWithMarkers> polylinesWithMarkers = [];
-        final bool isSelected = itinerary == selectedItinerary;
-        final bool showOnlySelected = selectedItinerary.isOnlyShowItinerary;
-        if (!showOnlySelected || isSelected) {
-          final List<PlanItineraryLeg> compressedLegs = itinerary.compressLegs;
-          for (int i = 0; i < compressedLegs.length; i++) {
-            final PlanItineraryLeg leg = compressedLegs[i];
-            // Polyline
-            final List<LatLng> points = leg.accumulatedPoints.isNotEmpty
-                ? leg.accumulatedPoints
-                : decodePolyline(leg.points);
-            final Color color = isSelected
-                ? leg.transportMode == TransportMode.bicycle &&
-                        leg.fromPlace.bikeRentalStation != null
-                    ? getBikeRentalNetwork(
-                            leg.fromPlace.bikeRentalStation.networks[0])
-                        .color
-                    : (leg?.route?.color != null
-                        ? Color(int.tryParse("0xFF${leg.route.color}"))
-                        : leg.transportMode.color)
-                : Colors.grey;
-            final Polyline polyline = Polyline(
-              points: points,
-              color: color,
-              strokeWidth: isSelected ? 6.0 : 3.0,
-              isDotted: leg.transportMode == TransportMode.walk,
-            );
-
-            // Transfer marker
-            if (isSelected &&
-                i < compressedLegs.length - 1 &&
-                polyline.points.isNotEmpty) {
-              markers.add(
-                buildTransferMarker(
-                  polyline.points[polyline.points.length - 1],
-                ),
-              );
-            }
-
-            // Bus marker
-            if (showTransportMarker &&
-                leg.transportMode != TransportMode.walk &&
-                leg.transportMode != TransportMode.bicycle &&
-                leg.transportMode != TransportMode.car) {
-              markers.add(
-                buildBusMarker(
-                  midPointForPolyline(polyline),
-                  isSelected
-                      ? (leg?.route?.color != null
-                          ? Color(int.tryParse("0xFF${leg.route.color}"))
-                          : leg.transportMode.backgroundColor)
-                      : Colors.grey,
-                  leg,
-                  icon: (leg?.route?.type ?? 0) == 715
-                      ? onDemandTaxiSvg(color: 'FFFFFF')
-                      : null,
-                  onTap: () => onTap(itinerary),
-                ),
-              );
-            }
-            polylinesWithMarkers.add(PolylineWithMarkers(polyline, markers));
-          }
-        }
-        itineraries.addAll({itinerary: polylinesWithMarkers});
-      }
-    }
-    return itineraries;
-  }
-}
-
-class PolylineWithMarkers {
-  PolylineWithMarkers(this.polyline, this.markers);
-
-  final Polyline polyline;
-  final List<Marker> markers;
 }
